Handle piece with no revisions
Right now, a user can go to the create a new piece page, which actually
creates a piece, and then leave it, which means they haven't created a
revision yet. The pieces index now only grabs pieces that have at least
a single revision, so the index can be displayed properly. This might
not be the ideal work-flow, but the index doesn't break becuase of it.