Show history link on Pieces#show; include specs
Instead of showing "all versions" or "all drafts", Pieces#show has a
"history" link in the piece bar. Add specs for this.
Add history to pieces routes for route and helper methods.
Refactor the piece bar versions partial to be specific to versions.
Add in a piece bar drafts partial to have functionality akin to
versions.