Issue #47 - Stop chance of duplicate draft numbers
Add Piece#create_draft to handle creating drafts for a piece.
Refactor DraftsController#create to use this method.
A new draft's number is based on the last draft's number, instead of
being based on the number of drafts that piece has. This removes the
possibility of duplicate draft numbers if an older draft is deleted.